api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
sqlalchemy
alembic
psycopg2-binary
boto3
python-multipart
confluent-kafka
watchdog
watchdog[watchmedo]
pymupdf
openai
tiktoken
pypdf
marshmallow-sqlalchemy
click
bcrypt
elasticsearch
requests
pip-review
sentry-sdk
python-jose[cryptography]
supertokens-python
pdf2image
asyncio
asyncpg
psycopg2-binary
aiobotocore
aioboto3==12.4.0
urllib3
aiofiles
transformers
async-timeout
pydub
pylint
black
isort
mypy
radon
bandit
docformatter
autoflake
types-aiofiles
types-requests
pytest